+++ src/compas/robots/model/joint.py
@@ -608,7 +608,10 @@ class Joint(Base):
         self.attr = kwargs
         self.child_link = None
         self.position = 0
-        self.current_origin = Origin(self.origin.point, self.origin.xaxis, self.origin.yaxis)
+        # The following are world-relative frames representing the origin and the axis, which change with
+        # the joint state, while `origin` and `axis` above are parent-relative and static.
+        self.current_origin = self.origin.copy()
+        self.current_axis = self.axis.copy()
 
     def get_urdf_element(self):
         attributes = {
@@ -696,7 +699,7 @@ class Joint(Base):
         None
         """
         self.current_origin.transform(transformation)
-        self.axis.transform(transformation)
+        self.current_axis.transform(transformation)
 
     def _create(self, transformation):
         """Internal method to initialize the transformation tree.
@@ -711,7 +714,7 @@ class Joint(Base):
         None
         """
         self.current_origin.transform(transformation)
-        self.axis.transform(self.current_transformation)
+        self.current_axis.transform(self.current_transformation)
 
     def calculate_revolute_transformation(self, position):
         """Returns a transformation of a revolute joint.
@@ -751,7 +754,7 @@ class Joint(Base):
         :class:`Rotation`
             Transformation of type rotation for the continuous joint.
         """
-        return Rotation.from_axis_and_angle(self.axis.vector, position, self.current_origin.point)
+        return Rotation.from_axis_and_angle(self.current_axis.vector, position, self.current_origin.point)
 
     def calculate_prismatic_transformation(self, position):
         """Returns a transformation of a prismatic joint.
@@ -774,7 +777,7 @@ class Joint(Base):
             raise ValueError('Prismatic joints are required to define a limit')
 
         position = max(min(position, self.limit.upper), self.limit.lower)
-        return Translation.from_vector(self.axis.vector * position)
+        return Translation.from_vector(self.current_axis.vector * position)
 
     # does this ever happen?
     def calculate_fixed_transformation(self, position):